clean up more shared code
