+++ src/tests/EntityGraphQL.Tests/MutationTests/PeopleMutations.cs
@@ -105,11 +105,6 @@ namespace EntityGraphQL.Tests
         {
             return args.Decimal;
         }
-        [GraphQLMutation]
-        public bool NoArgsWithService(AgeService ageService)
-        {
-            return ageService != null;
-        }
 
         [GraphQLMutation]
         public Expression<Func<TestDataContext, Person>> AddPersonAdv(PeopleMutationsArgs args)
